requirements.txt
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.8.1
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cpm-kernels==1.0.11
cycler==0.11.0
d2l==1.0.0b0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
git-lfs==1.6
gitdb==4.0.10
GitPython==3.1.32
gpytorch==1.11
gradio==3.38.0
gradio_client==0.2.10
gym==0.21.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jaxtyping==0.2.20
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
latex2mathml==3.76.0
linear-operator==0.5.1
linkify-it-py==2.0.2
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdtex2html==1.2.0
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
onnx==1.14.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.2
overrides==7.4.0
packaging==23.1
paddle-bfloat==0.1.7
paddlepaddle==2.5.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydeck==0.8.0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sse-starlette==1.6.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch @ https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-win_amd64.whl#sha256=b663a4ee744d574095dbd612644de345944247c0605692309fd9f6c7ccdea022
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.1
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
